Construction, fixes

+++ Server/src/main/core/game/system/command/sets/DevelopmentCommandSet.kt
@@ -29,50 +29,18 @@ import java.io.FileWriter
 @Initializable
 class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
 
-    private val farmKitItems = arrayListOf(
-        Items.RAKE_5341,
-        Items.SPADE_952,
-        Items.SEED_DIBBER_5343,
-        Items.WATERING_CAN8_5340,
-        Items.SECATEURS_5329,
-        Items.GARDENING_TROWEL_5325
-    )
-    private val runeKitItems = arrayListOf(
-        Items.AIR_RUNE_556,
-        Items.EARTH_RUNE_557,
-        Items.FIRE_RUNE_554,
-        Items.WATER_RUNE_555,
-        Items.MIND_RUNE_558,
-        Items.BODY_RUNE_559,
-        Items.DEATH_RUNE_560,
-        Items.NATURE_RUNE_561,
-        Items.CHAOS_RUNE_562,
-        Items.LAW_RUNE_563,
-        Items.COSMIC_RUNE_564,
-        Items.BLOOD_RUNE_565,
-        Items.SOUL_RUNE_566,
-        Items.ASTRAL_RUNE_9075
-    )
+    private val farmKitItems = arrayListOf(Items.RAKE_5341, Items.SPADE_952, Items.SEED_DIBBER_5343, Items.WATERING_CAN8_5340, Items.SECATEURS_5329, Items.GARDENING_TROWEL_5325)
+    private val runeKitItems = arrayListOf(Items.AIR_RUNE_556, Items.EARTH_RUNE_557, Items.FIRE_RUNE_554, Items.WATER_RUNE_555, Items.MIND_RUNE_558, Items.BODY_RUNE_559, Items.DEATH_RUNE_560, Items.NATURE_RUNE_561, Items.CHAOS_RUNE_562, Items.LAW_RUNE_563, Items.COSMIC_RUNE_564, Items.BLOOD_RUNE_565, Items.SOUL_RUNE_566, Items.ASTRAL_RUNE_9075)
 
     override fun defineCommands() {
 
-        define(
-            name = "farmkit",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Provides a kit of various farming equipment."
-        ) { player, _ ->
+        define(name = "farmkit", privilege = Privilege.ADMIN, usage = "", description = "Provides a kit of various farming equipment.") { player, _ ->
             for (item in farmKitItems) {
                 player.inventory.add(Item(item))
             }
         }
 
-        define(
-            name = "cs2",
-            privilege = Privilege.ADMIN,
-            usage = "::cs2 id args",
-            description = "Allows you to call arbitrary cs2 scripts during runtime"
-        ) { player, args ->
+        define(name = "cs2", privilege = Privilege.ADMIN, usage = "::cs2 id args", description = "Allows you to call arbitrary cs2 scripts during runtime") { player, args ->
             var scriptArgs = ArrayList<Any>()
             if (args.size == 2) {
                 runcs2(player, args[1].toIntOrNull() ?: return@define)
@@ -89,7 +57,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             }
         }
 
-        define("cleardiary", Privilege.ADMIN) { player, _ ->
+        define(name = "cleardiary", privilege = Privilege.ADMIN) { player, _ ->
             for (type in DiaryType.values()) {
                 val diary = player.achievementDiaryManager.getDiary(type)
                 if (diary != null) {
@@ -103,7 +71,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             sendMessage(player, "All achievement diaries cleared successfully.")
         }
 
-        define("clearjob", Privilege.ADMIN) { player, _ ->
+        define(name = "clearjob", privilege = Privilege.ADMIN) { player, _ ->
             val playerJobManager = JobManager.getInstance(player)
             playerJobManager.job = null
             playerJobManager.jobAmount = -1
@@ -116,12 +84,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             Prints current Region ID.
          */
 
-        define(
-            name = "region",
-            privilege = Privilege.STANDARD,
-            usage = "::region",
-            description = "Prints your current Region ID."
-        ) { player, _ ->
+        define(name = "region", privilege = Privilege.STANDARD, usage = "::region", description = "Prints your current Region ID.") { player, _ ->
             sendMessage(player, "Region ID: ${player.viewport.region.regionId}")
         }
 
@@ -129,12 +92,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             Swaps the spellbooks.
          */
 
-        define(
-            name = "spellbook",
-            privilege = Privilege.ADMIN,
-            usage = "::spellbook <lt>book ID<gt> (0 = MODERN, 1 = ANCIENTS, 2 = LUNARS)",
-            description = "Swaps your spellbook to the given book ID."
-        ) { player, args ->
+        define(name = "spellbook", privilege = Privilege.ADMIN, usage = "::spellbook <lt>book ID<gt> (0 = MODERN, 1 = ANCIENTS, 2 = LUNARS)", description = "Swaps your spellbook to the given book ID.") { player, args ->
             if (args.size < 2) {
                 reject(player, "Usage: ::spellbook [int]. 0 = MODERN, 1 = ANCIENTS, 2 = LUNARS")
             }
@@ -147,35 +105,25 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             Kill yourself.
          */
 
-        define(
-            name = "killme",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Does exactly what it says on the tin."
-        ) { player, _ ->
+        define(name = "killme", privilege = Privilege.ADMIN, usage = "", description = "Does exactly what it says on the tin.") { player, _ ->
             player.impactHandler.manualHit(player, player.skills.lifepoints, HitsplatType.NORMAL)
         }
 
-        define("struct") { _, args ->
+        define(name = "struct") { _, args ->
             val mapId = args[1].toIntOrNull() ?: return@define
 
             val def = Struct.get(mapId)
             log(this::class.java, Log.FINE, def.toString())
         }
 
-        define("datamap") { _, args ->
+        define(name = "datamap") { _, args ->
             val mapId = args[1].toIntOrNull() ?: return@define
 
             val def = DataMap.get(mapId)
             log(this::class.java, Log.FINE, def.toString())
         }
 
-        define(
-            name = "dumpstructs",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Dumps all the cache structs to structs.txt"
-        ) { _, _ ->
+        define(name = "dumpstructs", privilege = Privilege.ADMIN, usage = "", description = "Dumps all the cache structs to structs.txt") { _, _ ->
             val dump = File("structs.txt")
             val writer = BufferedWriter(FileWriter(dump))
             val index = Cache.getIndexes()[2]
@@ -193,12 +141,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             writer.close()
         }
 
-        define(
-            name = "dumpdatamaps",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Dumps all the cache data maps to datamaps.txt"
-        ) { _, _ ->
+        define(name = "dumpdatamaps", privilege = Privilege.ADMIN, usage = "", description = "Dumps all the cache data maps to datamaps.txt") { _, _ ->
             val index = Cache.getIndexes()[17]
             val containers = index.information.containersIndexes
 
@@ -221,12 +164,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             writer.close()
         }
 
-        define(
-            name = "rolldrops",
-            privilege = Privilege.ADMIN,
-            usage = "::rolldrops <lt>NPC ID<gt> <lt>AMOUNT<gt>",
-            description = "Rolls the given NPC drop table AMOUNT times."
-        ) { player: Player, args: Array<String> ->
+        define(name = "rolldrops", privilege = Privilege.ADMIN, usage = "::rolldrops <lt>NPC ID<gt> <lt>AMOUNT<gt>", description = "Rolls the given NPC drop table AMOUNT times.") { player: Player, args: Array<String> ->
             if (args.size < 2) {
                 reject(player, "Usage: ::rolldrops npcid amount")
             }
@@ -241,12 +179,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             container.open(player)
         }
 
-        define(
-            name = "varbits",
-            privilege = Privilege.ADMIN,
-            usage = "::varbits <lt>Varp ID<gt>",
-            description = "Lists all the varbits assigned to the given varp."
-        ) { player, args ->
+        define(name = "varbits", privilege = Privilege.ADMIN, usage = "::varbits <lt>Varp ID<gt>", description = "Lists all the varbits assigned to the given varp.") { player, args ->
             if (args.size < 2)
                 reject(player, "Usage: ::varbits varpIndex")
 
@@ -263,16 +196,11 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             }
         }
 
-        define("testpacket") { player, _ ->
+        define(name = "testpacket") { player, _ ->
             PacketWriteQueue.write(ResetInterface(), PlayerContext(player))
         }
 
-        define(
-            name = "npcsearch",
-            privilege = Privilege.STANDARD,
-            usage = "npcsearch name",
-            description = "Searches for NPCs that match the name either in main or children."
-        ) { player, strings ->
+        define(name = "npcsearch", privilege = Privilege.STANDARD, usage = "npcsearch name", description = "Searches for NPCs that match the name either in main or children.") { player, strings ->
             val name = strings.slice(1 until strings.size).joinToString(" ").lowercase()
             for (id in 0 until 9000) {
                 val def = NPCDefinition.forId(id)
@@ -291,7 +219,7 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             }
         }
 
-        define("itemsearch") { player, args ->
+        define(name = "itemsearch") { player, args ->
             val itemName = args.copyOfRange(1, args.size).joinToString(" ").lowercase()
             for (i in 0 until 15000) {
                 val name = getItemName(i).lowercase()
@@ -300,95 +228,45 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             }
         }
 
-        define(
-            name = "runekit",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Gives 1k of each Rune type"
-        ) { player, _ ->
+        define(name = "runekit", privilege = Privilege.ADMIN, usage = "", description = "Gives 1k of each Rune type") { player, _ ->
             for (item in runeKitItems) {
                 addItem(player, item, 1000)
             }
         }
 
-        define(
-            name = "drawchunks",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Draws the border of the chunk you're standing in"
-        ) { player, _ ->
+        define(name = "drawchunks", privilege = Privilege.ADMIN, usage = "", description = "Draws the border of the chunk you're standing in") { player, _ ->
             setAttribute(player, "chunkdraw", !getAttribute(player, "chunkdraw", false))
         }
 
-        define(
-            name = "drawclipping",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Draws the clipping flags of the region you're standing in"
-        ) { player, _ ->
+        define(name = "drawclipping", privilege = Privilege.ADMIN, usage = "", description = "Draws the clipping flags of the region you're standing in") { player, _ ->
             setAttribute(player, "clippingdraw", !getAttribute(player, "clippingdraw", false))
         }
 
-        define(
-            name = "drawregions",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "DRaws the border of the region you're standing in"
-        ) { player, _ ->
+        define(name = "drawregions", privilege = Privilege.ADMIN, usage = "", description = "DRaws the border of the region you're standing in") { player, _ ->
             setAttribute(player, "regiondraw", !getAttribute(player, "regiondraw", false))
         }
 
-        define(
-            name = "drawroute",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Visualizes the path your player is taking"
-        ) { player, _ ->
+        define(name = "drawroute", privilege = Privilege.ADMIN, usage = "", description = "Visualizes the path your player is taking") { player, _ ->
             setAttribute(player, "routedraw", !getAttribute(player, "routedraw", false))
         }
 
-        define(
-            name = "fmstart",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, _ ->
+        define(name = "fmstart", privilege = Privilege.ADMIN, usage = "", description = "") { player, _ ->
             setAttribute(player, "fmstart", Location.create(player.location))
         }
 
-        define(
-            name = "fmend",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, _ ->
+        define(name = "fmend", privilege = Privilege.ADMIN, usage = "", description = "") { player, _ ->
             setAttribute(player, "fmend", Location.create(player.location))
         }
 
-        define(
-            name = "fmspeed",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, args ->
+        define(name = "fmspeed", privilege = Privilege.ADMIN, usage = "", description = "") { player, args ->
             setAttribute(player, "fmspeed", args[1].toIntOrNull() ?: 10)
         }
 
-        define(
-            name = "fmspeedend",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, args ->
+        define(name = "fmspeedend", privilege = Privilege.ADMIN, usage = "", description = "") { player, args ->
             setAttribute(player, "fmspeedend", args[1].toIntOrNull() ?: 10)
         }
 
-        define(
-            name = "testfm",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, _ ->
+        define(name = "testfm", privilege = Privilege.ADMIN, usage = "", description = "") { player, _ ->
             val start = getAttribute(player, "fmstart", Location.create(player.location))
             val end = getAttribute(player, "fmend", Location.create(player.location))
             val speed = getAttribute(player, "fmspeed", 10)
@@ -397,42 +275,22 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             forceMove(player, start, end, speed, speedEnd, anim = ani)
         }
 
-        define(
-            name = "fmanim",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = ""
-        ) { player, args ->
+        define(name = "fmanim", privilege = Privilege.ADMIN, usage = "", description = "") { player, args ->
             setAttribute(player, "fmanim", args[1].toIntOrNull() ?: -1)
         }
 
-        define(
-            name = "drawintersect",
-            privilege = Privilege.ADMIN,
-            usage = "",
-            description = "Visualizes the predicted intersection point with an NPC"
-        ) { player, _ ->
+        define(name = "drawintersect", privilege = Privilege.ADMIN, usage = "", description = "Visualizes the predicted intersection point with an NPC") { player, _ ->
             setAttribute(player, "draw-intersect", !getAttribute(player, "draw-intersect", false))
         }
 
-        define(
-            name = "expression",
-            privilege = Privilege.ADMIN,
-            usage = "::expression id",
-            description = "Visualizes chathead animations from ID."
-        ) { player, args ->
+        define(name = "expression", privilege = Privilege.ADMIN, usage = "::expression id", description = "Visualizes chathead animations from ID.") { player, args ->
             if (args.size != 2)
                 reject(player, "Usage: ::expression id")
             val id = args[1].toIntOrNull() ?: 9804
             player.dialogueInterpreter.sendDialogues(player, id, "Expression ID: $id")
         }
 
-        define(
-            name = "timers",
-            privilege = Privilege.ADMIN,
-            usage = "::timers",
-            description = "Print out timers"
-        ) { player, _ ->
+        define(name = "timers", privilege = Privilege.ADMIN, usage = "::timers", description = "Print out timers") { player, _ ->
             player.sendMessage("Active timers:")
             for (timer in player.timers.activeTimers) {
                 player.sendMessage("  ${timer.identifier} ${timer.nextExecution}")
@@ -441,6 +299,16 @@ class DevelopmentCommandSet : CommandSet(Privilege.ADMIN) {
             for (timer in player.timers.newTimers) {
                 player.sendMessage("  ${timer.identifier}")
             }
+        }
+
+        define(name = "overlay", privilege = Privilege.ADMIN, usage = "::overlay <lt>Overlay ID<gt>") { player, args ->
+            val overlayInt = args[1].toInt()
+            openOverlay(player,overlayInt)
+        }
+
+        define(name = "interface", privilege = Privilege.ADMIN, usage = "::interface <lt>Interface ID<gt>") { player, args ->
+            val interfaceInt = args[1].toInt()
+            openInterface(player,interfaceInt)
         }
     }
 }
